requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
nest_asyncio>=1.6.0
psycopg2-binary>=2.9.0
databricks-sdk>=0.81.0
faster-whisper>=0.9.0
torch>=2.0.0
numpy>=1.24.0
soundfile>=0.12.0
gradio>=4.0.0
python-multipart>=0.0.9
